feat: implementa EstoqueItemService

+++ src/main/java/com/gestaooserp/dev/repository/EstoqueItemRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.EstoqueItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstoqueItemRepository extends JpaRepository<EstoqueItem,Integer> {
+public interface EstoqueItemRepository extends JpaRepository<EstoqueItem,Long> {
 }
